requirements.txt
# FastAPI backend dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
sentry-sdk>=2.0.0
sqlalchemy>=2.0.30
python-dotenv>=1.0.1
redis>=5.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
prometheus-client>=0.19.0
structlog>=24.1.0

# Supabase integration
supabase>=1.0.0
httpx>=0.27.0

# Cloudflare integration
httpx>=0.27.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# Google Drive integration for Ollama models
google-api-python-client>=2.127.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.2.0

# Secrets adapter dependencies
hvac>=1.2.0
aiohttp>=3.9.0

# Additional dependencies from main requirements
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
pydantic>=2.0.0
cryptography>=42.0.0
email-validator>=2.3.0
bleach>=6.1.0
openai>=1.12.0
anthropic>=0.21.3
psutil>=5.9.0
celery[redis]>=5.3.0
flower>=2.0.0
APScheduler>=3.10.0

# Sandbox dependencies
rq>=1.15.0
docker>=7.0.0
boto3>=1.34.0